...G...PS. [ZBX-13660] fixed data types to avoid runtime warnings like 'control.c:163:13: runtime error: left shift of 32768 by 16 places cannot be represented in type 'int' and compiler warnings

+++ src/libs/zbxnix/control.c
@@ -19,7 +19,7 @@
 
 #include "control.h"
 
-static int	parse_log_level_options(const char *opt, size_t len, int *scope, int *data)
+static int	parse_log_level_options(const char *opt, size_t len, unsigned int *scope, unsigned int *data)
 {
 	unsigned short	num = 0;
 	const char	*rtc_options;
@@ -99,7 +99,7 @@ static int	parse_log_level_options(const char *opt, size_t len, int *scope, int 
 
 		zbx_free(proc_name);
 
-		*scope = ZBX_RTC_LOG_SCOPE_PROC | proc_type;
+		*scope = ZBX_RTC_LOG_SCOPE_PROC | (unsigned int)proc_type;
 		*data = num;
 	}
 
@@ -124,7 +124,7 @@ static int	parse_log_level_options(const char *opt, size_t len, int *scope, int 
  ******************************************************************************/
 int	parse_rtc_options(const char *opt, unsigned char program_type, int *message)
 {
-	int	scope, data, command;
+	unsigned int	scope, data, command;
 
 	if (0 == strncmp(opt, ZBX_LOG_LEVEL_INCREASE, ZBX_CONST_STRLEN(ZBX_LOG_LEVEL_INCREASE)))
 	{
@@ -160,7 +160,7 @@ int	parse_rtc_options(const char *opt, unsigned char program_type, int *message)
 		return FAIL;
 	}
 
-	*message = ZBX_RTC_MAKE_MESSAGE(command, scope, data);
+	*message = (int)ZBX_RTC_MAKE_MESSAGE(command, scope, data);
 
 	return SUCCEED;
 }
